+++ apps/extension/src/app/store/app-permissions/app-permissions.selectors.ts
@@ -1,0 +1,8 @@
+import { RootState } from '..';
+import { appPermissionsAdapter } from './app-permissions.slice';
+
+function selectAppPermissionsSlice(state: RootState) {
+  return state.appPermissions;
+}
+export const appPermissionsSelectors =
+  appPermissionsAdapter.getSelectors<RootState>(selectAppPermissionsSlice);