+++ apps/mobile/CHANGELOG.md
@@ -23,6 +23,50 @@
 * axios cve ([b145cd5](https://github.com/leather-io/mono/commit/b145cd5fd055201b18304b5e2b92635d56086915))
 * **mobile:** help header ([764af25](https://github.com/leather-io/mono/commit/764af257655ecb585d6805c5350c3d5fffe7a56a))
 
+## [2.112.1](https://github.com/leather-io/mono/compare/@leather.io/mobile-v2.112.0...@leather.io/mobile-v2.112.1) (2026-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/bitcoin bumped to 0.39.2
+    * @leather.io/features bumped to 1.5.2
+    * @leather.io/provider bumped to 1.6.31
+    * @leather.io/queries bumped to 0.14.0
+    * @leather.io/query bumped to 2.51.2
+    * @leather.io/rpc bumped to 2.24.1
+    * @leather.io/services bumped to 1.60.1
+    * @leather.io/stacks bumped to 1.22.0
+    * @leather.io/ui bumped to 1.114.2
+
+## [2.112.0](https://github.com/leather-io/mono/compare/@leather.io/mobile-v2.111.0...@leather.io/mobile-v2.112.0) (2026-07-15)
+
+
+### Features
+
+* **extension:** add wallet fingerprint to getAddresses ([#2499](https://github.com/leather-io/mono/issues/2499)) ([8e13787](https://github.com/leather-io/mono/commit/8e13787b37de92b334ea98dbb371ac03c622848f))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/analytics bumped to 3.21.1
+    * @leather.io/bitcoin bumped to 0.39.1
+    * @leather.io/constants bumped to 0.37.1
+    * @leather.io/crypto bumped to 1.13.2
+    * @leather.io/features bumped to 1.5.1
+    * @leather.io/models bumped to 0.60.0
+    * @leather.io/provider bumped to 1.6.30
+    * @leather.io/queries bumped to 0.13.0
+    * @leather.io/query bumped to 2.51.1
+    * @leather.io/rpc bumped to 2.24.0
+    * @leather.io/services bumped to 1.60.0
+    * @leather.io/stacks bumped to 1.21.1
+    * @leather.io/ui bumped to 1.114.1
+    * @leather.io/utils bumped to 0.52.2
+
 ## [2.111.0](https://github.com/leather-io/mono/compare/@leather.io/mobile-v2.110.0...@leather.io/mobile-v2.111.0) (2026-07-10)