+++ packages/bitcoin/CHANGELOG.md
@@ -939,6 +939,29 @@
     * @leather.io/prettier-config bumped to 0.9.0
     * @leather.io/rpc bumped to 2.20.17
 
+## [0.39.2](https://github.com/leather-io/mono/compare/@leather.io/bitcoin-v0.39.1...@leather.io/bitcoin-v0.39.2) (2026-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @leather.io/rpc bumped to 2.24.1
+
+## [0.39.1](https://github.com/leather-io/mono/compare/@leather.io/bitcoin-v0.39.0...@leather.io/bitcoin-v0.39.1) (2026-07-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @leather.io/constants bumped to 0.37.1
+    * @leather.io/crypto bumped to 1.13.2
+    * @leather.io/models bumped to 0.60.0
+    * @leather.io/utils bumped to 0.52.2
+  * devDependencies
+    * @leather.io/rpc bumped to 2.24.0
+
 ## [0.39.0](https://github.com/leather-io/mono/compare/@leather.io/bitcoin-v0.38.1...@leather.io/bitcoin-v0.39.0) (2026-07-10)